+++ Documentation/MVVMDemo/MVVMDemo/Views/AppView/AppView.swift
@@ -74,6 +74,8 @@ struct AppView: View {
 struct AppView_Previews: PreviewProvider {
     static var previews: some View {
         AppView().environment(\.playerRepository, .empty())
+            .previewDisplayName("Database Initially Empty")
         AppView().environment(\.playerRepository, .populated())
+            .previewDisplayName("Database Initially Populated")
     }
 }
