+++ Documentation/QueryDemo/QueryDemo/Views/DatabaseButtons.swift
@@ -22,19 +22,19 @@ struct CreatePlayerButton: View {
 /// A helper button that deletes players in the database
 struct DeletePlayersButton: View {
     private enum Mode {
-        case deleteAfter
-        case deleteBefore
+        case delete
+        case deleteAfter(() -> Void)
+        case deleteBefore(() -> Void)
     }
     
     @Environment(\.playerRepository) private var playerRepository
     private var titleKey: LocalizedStringKey
-    private var action: (() -> Void)?
     private var mode: Mode
     
     /// Creates a button that simply deletes players.
     init(_ titleKey: LocalizedStringKey) {
         self.titleKey = titleKey
-        self.mode = .deleteBefore
+        self.mode = .delete
     }
     
     /// Creates a button that deletes players soon after performing `action`.
@@ -43,8 +43,7 @@ struct DeletePlayersButton: View {
         after action: @escaping () -> Void)
     {
         self.titleKey = titleKey
-        self.action = action
-        self.mode = .deleteAfter
+        self.mode = .deleteAfter(action)
     }
     
     /// Creates a button that deletes players immediately after performing `action`.
@@ -53,22 +52,25 @@ struct DeletePlayersButton: View {
         before action: @escaping () -> Void)
     {
         self.titleKey = titleKey
-        self.action = action
-        self.mode = .deleteBefore
+        self.mode = .deleteBefore(action)
     }
     
     var body: some View {
         Button {
             switch mode {
-            case .deleteAfter:
-                action?()
-                DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                    _ = try! playerRepository.deleteAllPlayer()
+            case .delete:
+                _ = try! playerRepository.deleteAllPlayer()
+                
+            case let .deleteAfter(action):
+                action()
+                Task {
+                    try await Task.sleep(nanoseconds: 100_000_000)
+                    try playerRepository.deleteAllPlayer()
                 }
                 
-            case .deleteBefore:
+            case let .deleteBefore(action):
                 _ = try! playerRepository.deleteAllPlayer()
-                action?()
+                action()
             }
         } label: {
             Label(titleKey, systemImage: "trash")
@@ -76,8 +78,9 @@ struct DeletePlayersButton: View {
     }
 }
 
+// For tracking the player count in the preview
 import GRDB
-import GRDBQuery // For tracking the player count in the preview
+import GRDBQuery
 
 struct DatabaseButtons_Previews: PreviewProvider {
     struct PlayerCountRequest: Queryable {
